Automatically provide global using
The new C# 10 feature lights up automatically via the provided item group with the `Import` element.

There's talk of a future opt-out (or opt-in?) MSBuild property to control these global usings, but the provided approach will work and integrate seamlessly with whatever the project/SDK configures.

For reference, see https://github.com/dotnet/aspnetcore/issues/32451.

Note we provide the "legacy" MSBuild xmlns attribute so as not to break downlevel IDEs or legacy project files that happen to reference this package.

Fixes #515