Custom stack trace and method name resolver for Exception formatting
Newer versions of .NET clean up the exception messages internally for things like async and the such. They clean the stack trace and also tweak the names. This gives a pretty big difference between our output and the default ToString. 

This allows the user to do their own tweaking to the method names and stack frames before the formatting. Or even better, use something like Ben.Demystifer to do the heavy lifting for them.