Performance issue solved
When printing tens of thousands of lines, there is a performance issue. The TableRowCollection -> Add method was being called millions of times.
Once  it is called, it returns the index  of the added element.
As the element is added to a List (new List<TableRow>, I believe it is not needed to use IndexOf because it will be added always at the last position.
Replacing the IndexOf call by a Count - 1 improves the performance in this scenario.
Related issue: #975