Better use of DynamicallyAccessedMemberTypes and UnconditionalSuppressMessage
Tried to rip out UnconditionalSuppressMessage where types could be annotated properly. Also tried to get DynamicallyAccessedMemberTypes to better represent what
members were being accessed - almost always it is a public constructor and public properties.

Part of this exercise showed the need for a new CommandApp for DefaultCommands that accepts a parameter. Right now it lives side-by-side with the original CommandApp<T>