Condition `PolySharp` for only `netstandard2.0`
Context: https://github.com/dotnet/sdk/pull/51430
Context: https://github.com/dotnet/source-build-reference-packages/pull/1447

We would like to use Spectre.Console for the `dotnet` CLI, and .NET has
to be able to be completely built from source.

It would help us to have `PolySharp` only included when building for
`netstandard2.0`, as that is the only target framework that needs the
polyfills.

.NET would probably only use the latest target framework.

Thanks!